refactor: remove `Suspense` in `SideMenuLayout` and unnecessary layout files
