+++ src/app/(side-menu-layout)/layout.tsx
@@ -1,0 +1,17 @@
+import { fetchRecentCopipes } from "@/db/server/copipes"
+import { fetchTags } from "@/db/server/tags"
+import SideMenuLayout from "@/modules/layouts/sideMenuLayout"
+import { PropsWithChildren } from "react"
+
+interface LayoutProps extends PropsWithChildren {}
+
+export default async function Layout({ children }: LayoutProps) {
+  const copipes = await fetchRecentCopipes()
+  const tags = await fetchTags()
+
+  return (
+    <SideMenuLayout copipes={copipes} tags={tags}>
+      {children}
+    </SideMenuLayout>
+  )
+}
