--- src/app/(side-menu-layout)/loading.tsx
@@ -1,9 +1,0 @@
-import { Center, Loading as UILoading } from "@yamada-ui/react"
-
-export default function Loading() {
-  return (
-    <Center h="3xs" w="full">
-      <UILoading fontSize="2xl" />
-    </Center>
-  )
-}
