--- src/app/(side-menu-layout)/search/layout.tsx
@@ -1,9 +1,0 @@
-import { ReactNode } from "react"
-
-export const metadata = {
-  title: "検索",
-}
-
-export default function layout({ children }: { children: ReactNode }) {
-  return <>{children}</>
-}
