+++ src/app/(side-menu-layout)/search/page.tsx
@@ -1,8 +1,12 @@
-import SideMenuLayout from "@/modules/layouts/sideMenuLayout"
+import { Metadata } from "next/types"
 import { Suspense } from "react"
 
 import { SearchPageLoading } from "./page-loading"
 import { SearchPageTemplate } from "./page-template"
+
+export const metadata: Metadata = {
+  title: "検索",
+}
 
 export default function page({
   searchParams,
@@ -10,10 +14,8 @@ export default function page({
   searchParams: Promise<Record<string, string | string[] | undefined>>
 }) {
   return (
-    <SideMenuLayout>
-      <Suspense fallback={<SearchPageLoading />}>
-        <SearchPageTemplate searchParams={searchParams} />
-      </Suspense>
-    </SideMenuLayout>
+    <Suspense fallback={<SearchPageLoading />}>
+      <SearchPageTemplate searchParams={searchParams} />
+    </Suspense>
   )
 }
