--- src/app/(side-menu-layout)/tag/[tagId]/layout.tsx
@@ -1,10 +1,0 @@
-import { Metadata } from "next"
-import { ReactNode } from "react"
-
-export const metadata: Metadata = {
-  title: "タグ",
-}
-
-export default function layout({ children }: { children: ReactNode }) {
-  return <>{children}</>
-}
