+++ src/modules/layouts/sideMenuLayout.tsx
@@ -1,14 +1,21 @@
+import { FetchRecentCopipesReturn } from "@/db/server/copipes"
+import { FetchTagsReturn } from "@/db/server/tags"
 import RecentPostsCard from "@/modules/recentPostCard/recentPostCard"
 import TagListCard from "@/modules/tagListCard/tagListCard"
 import { GridItem, SimpleGrid, VStack } from "@yamada-ui/react"
-import { PropsWithChildren, Suspense } from "react"
+import { PropsWithChildren } from "react"
 
-import LoadingRecentPostsCard from "../recentPostCard/loading"
-import LoadingTagListCard from "../tagListCard/loading"
 
-interface SideMenuLayoutProps extends PropsWithChildren {}
+interface SideMenuLayoutProps extends PropsWithChildren {
+  copipes: Awaited<FetchRecentCopipesReturn>
+  tags: Awaited<FetchTagsReturn>
+}
 
-export default function SideMenuLayout({ children }: SideMenuLayoutProps) {
+export default function SideMenuLayout({
+  children,
+  copipes,
+  tags,
+}: SideMenuLayoutProps) {
   return (
     <SimpleGrid apply="gridStyle.grid" as="main" columns={4}>
       <GridItem apply="gridStyle.item" colSpan={{ base: 3, md: 4 }}>
@@ -21,13 +28,9 @@ export default function SideMenuLayout({ children }: SideMenuLayoutProps) {
         display={{ base: "block", md: "none" }}
       >
         <VStack>
-          <Suspense fallback={<LoadingTagListCard />}>
-            <TagListCard />
-          </Suspense>
+          <TagListCard tags={tags} />
 
-          <Suspense fallback={<LoadingRecentPostsCard />}>
-            <RecentPostsCard />
-          </Suspense>
+          <RecentPostsCard copipes={copipes} />
         </VStack>
       </GridItem>
     </SimpleGrid>
